ensure not create favorite location with invalid lat

+++ api/GloboClima.Api/Inputs/CreateFavoriteLocationInput.cs
@@ -4,8 +4,8 @@ namespace GloboClima.Api.Inputs;
 
 public class CreateFavoriteLocationInput
 {
-    public required double Lat { get; init; }
-    public required double Lon { get; init; }
+    public required int Lat { get; init; }
+    public required int Lon { get; init; }
 
     public FavoriteLocation GetFavoriteLocation()
         => new() { Lat = Lat, Lon = Lon };
